Centralize worker-count parsing to keep setup examples concise
